Read cron_interval_minutes from env var if not defined in the config yaml.
