+++ config.yml.sample
@@ -39,9 +39,9 @@ queries: [
     api_ver: "58.0"
   },
   {
-    query: "SELECT EventName, EventType, UsageType, Client, Value, StartDate, EndDate FROM PlatformEventUsageMetric WHERE TimeSegment='FifteenMinutes' AND StartDate >= {start_date} AND EndDate <= {now}",
+    query: "SELECT EventName, EventType, UsageType, Client, Value, StartDate, EndDate FROM PlatformEventUsageMetric WHERE TimeSegment='FifteenMinutes' AND StartDate >= {start_date} AND EndDate <= {end_date}",
     env: {
-      now: "now()",
+      end_date: "now()",
       start_date: "now(timedelta(minutes=-60))"
     },
     api_ver: "58.0",
